requirements.txt
boto3
neo4j
python-dotenv
google-genai
